pyproject.toml
[project]
name = "scifi-evals"
version = "0.1.0"
description = "Sci-Fi Ethical Dilemmas Evaluation System - A comprehensive system for evaluating how different LLMs respond to ethical dilemmas from classic science fiction literature"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.108.1",
    "python-dotenv>=1.1.1",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "sentence-transformers>=3",
    "typer>=0.12",
    "numpy>=2",
    "torch>=2.4",
    "orjson>=3.10",
    "pytest>=8.4.2",
]

[dependency-groups]
dev = [
    "pytest>=8",
]
